+++ ui-lib/src/main/java/co/electriccoin/zcash/di/DateSourceModule.kt
@@ -2,8 +2,6 @@ package co.electriccoin.zcash.di
 
 import co.electriccoin.zcash.ui.common.datasource.LocalAddressBookDataSource
 import co.electriccoin.zcash.ui.common.datasource.LocalAddressBookDataSourceImpl
-import co.electriccoin.zcash.ui.common.datasource.RemoteAddressBookDataSource
-import co.electriccoin.zcash.ui.common.datasource.RemoteAddressBookDataSourceImpl
 import org.koin.core.module.dsl.singleOf
 import org.koin.dsl.bind
 import org.koin.dsl.module
@@ -11,5 +9,5 @@ import org.koin.dsl.module
 val dataSourceModule =
     module {
         singleOf(::LocalAddressBookDataSourceImpl) bind LocalAddressBookDataSource::class
-        singleOf(::RemoteAddressBookDataSourceImpl) bind RemoteAddressBookDataSource::class
+        // singleOf(::RemoteAddressBookDataSourceImpl) bind RemoteAddressBookDataSource::class
     }